rename model file for admin user
