+++ front-end-admin/projeto_salvaguarda_admin/lib/services/auth/repository/mock_auth_repository.dart
@@ -1,7 +1,7 @@
 import 'dart:async';
 
 import 'package:projeto_salvaguarda_admin/services/auth/errors/auth_exceptions.dart';
-import 'package:projeto_salvaguarda_admin/services/auth/models/user.dart';
+import 'package:projeto_salvaguarda_admin/services/auth/models/user_admin.dart';
 import 'package:projeto_salvaguarda_admin/services/auth/repository/auth_stream_repository.dart';
 
 class MockAuthRepository implements AuthStreamRepository {
