+++ front-end-admin/projeto_salvaguarda_admin/lib/services/auth/store/auth_store.dart
@@ -1,5 +1,5 @@
 import 'package:flutter/material.dart';
-import 'package:projeto_salvaguarda_admin/services/auth/models/user.dart';
+import 'package:projeto_salvaguarda_admin/services/auth/models/user_admin.dart';
 import 'package:projeto_salvaguarda_admin/services/auth/repository/auth_stream_repository.dart';
 
 class AuthStore extends ValueNotifier<SalvaGuardasAdmin?> {
